Restructuring code so button listeners are in the fragments instead of main activities
