+++ app/src/main/res/layout/tutor_profile_edit.xml
@@ -37,14 +37,13 @@
             android:paddingBottom="24dp">
 
             <Button
-                android:id="@+id/change_profile_button"
+                android:id="@+id/change_picture_button"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="@string/change_picture_button_text"
                 android:textAllCaps="false"
                 android:layout_centerVertical="true"
-                android:layout_centerHorizontal="true"
-                android:onClick="changePicture"/>
+                android:layout_centerHorizontal="true"/>
 
             <TextView
                 android:id="@+id/profile_name_edit"
@@ -55,7 +54,7 @@
                 android:textColor="#fff"
                 android:textSize="24sp"
                 android:textStyle="bold"
-                android:layout_below="@+id/change_profile_button"
+                android:layout_below="@+id/change_picture_button"
                 android:layout_centerHorizontal="true"/>
 
         </RelativeLayout>
@@ -176,9 +175,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:id="@+id/save_profile_button"
-                android:text="@string/save_profile_button_text"
-                android:onClick="saveProfile"
-                />
+                android:text="@string/save_profile_button_text"/>
 
 
         </LinearLayout>
